refactor: service 하위 패키지 재구조
