+++ backend/src/main/java/org/donggle/backend/application/service/writing/WritingFacadeService.java
@@ -1,7 +1,7 @@
 package org.donggle.backend.application.service.writing;
 
 import lombok.RequiredArgsConstructor;
-import org.donggle.backend.application.NoConcurrentExecution;
+import org.donggle.backend.application.service.concurrent.NoConcurrentExecution;
 import org.donggle.backend.application.service.request.MarkdownUploadRequest;
 import org.donggle.backend.application.service.request.NotionUploadRequest;
 import org.donggle.backend.application.service.request.WritingModifyRequest;
