+++ backend/src/main/java/org/donggle/backend/ui/MemberController.java
@@ -1,7 +1,7 @@
 package org.donggle.backend.ui;
 
 import lombok.RequiredArgsConstructor;
-import org.donggle.backend.application.service.MemberService;
+import org.donggle.backend.application.service.member.MemberService;
 import org.donggle.backend.ui.common.AuthenticationPrincipal;
 import org.donggle.backend.ui.response.MemberPageResponse;
 import org.springframework.http.ResponseEntity;
