+++ backend/src/test/java/org/donggle/backend/application/service/TrashServiceTest.java
@@ -4,6 +4,7 @@ import jakarta.transaction.Transactional;
 import org.donggle.backend.application.repository.CategoryRepository;
 import org.donggle.backend.application.repository.MemberRepository;
 import org.donggle.backend.application.repository.WritingRepository;
+import org.donggle.backend.application.service.trash.TrashService;
 import org.donggle.backend.domain.category.Category;
 import org.donggle.backend.domain.member.Member;
 import org.donggle.backend.domain.writing.Title;
